Improve BPE testing.

+++ dstauffman/tests/test_bpe.py
@@ -598,10 +598,117 @@ class Test__double_dogleg(unittest.TestCase):
              self.gradient, self.grad_hessian_grad, self.x_bias, self.trust_radius)
 
 #%% _dogleg_search
-pass
+@patch('dstauffman.bpe.logger')
+class Test__dogleg_search(unittest.TestCase):
+    r"""
+    Tests the _dogleg_search function with the following cases:
+        TBD
+    """
+    def setUp(self):
+        dcs.bpe.logger.setLevel(logging.INFO)
+        time        = np.arange(251)
+        sim_params  = SimParams(time, magnitude=3.5, frequency=12, phase=180)
+        truth_time  = np.arange(-10, 201)
+        truth_data  = 5 * np.sin(2*np.pi*10*time/1000 + 90*np.pi/180)
+
+        self.opti_opts                = dcs.OptiOpts()
+        self.opti_opts.model_func     = sim_model
+        self.opti_opts.model_args     = {'sim_params': sim_params}
+        self.opti_opts.cost_func      = cost_wrapper
+        self.opti_opts.cost_args      = {'results_time': time, 'truth_time': truth_time, 'truth_data': truth_data}
+        self.opti_opts.get_param_func = get_parameter
+        self.opti_opts.set_param_func = set_parameter
+        self.opti_opts.output_folder  = ''
+        self.opti_opts.output_results = ''
+        self.opti_opts.params         = []
+
+        # Parameters to estimate
+        self.opti_opts.params.append(dcs.OptiParam('magnitude', best=2.5, min_=-10, max_=10, typical=5, minstep=0.01))
+        self.opti_opts.params.append(dcs.OptiParam('frequency', best=20, min_=1, max_=1000, typical=60, minstep=0.01))
+        self.opti_opts.params.append(dcs.OptiParam('phase', best=180, min_=0, max_=360, typical=100, minstep=0.1))
+
+        self.model_args = self.opti_opts.model_args
+
+        self.bpe_results = dcs.BpeResults()
+        self.cur_results = dcs.CurrentResults()
+
+        # initialize current results
+        (_, self.cur_results.innovs) = dcs.bpe._function_wrapper(self.opti_opts, self.bpe_results, self.model_args)
+        self.cur_results.trust_rad = self.opti_opts.trust_radius
+        self.cur_results.cost      = 0.5 * dcs.rss(self.cur_results.innovs, ignore_nans=True)
+        names = dcs.OptiParam.get_names(self.opti_opts.params)
+        self.cur_results.params    = self.opti_opts.get_param_func(names=names, **self.model_args)
+
+        # set relevant results variables
+        self.bpe_results.param_names  = [name.encode('utf-8') for name in names]
+        self.bpe_results.begin_params = self.cur_results.params.copy()
+        self.bpe_results.begin_innovs = self.cur_results.innovs.copy()
+        self.bpe_results.begin_cost   = self.cur_results.cost
+        self.bpe_results.costs.append(self.cur_results.cost)
+
+        self.delta_param = np.array([1, 2, 3])
+        self.gradient    = np.array([4, 5, 6])
+        self.hessian     = np.array([[5, 2, 1], [1, 2, 5], [3, 3, 3]])
+        self.jacobian    = np.random.rand(201, 3)
+        self.normalized  = False
+
+    def test_nominal(self, mock_logger):
+        dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+            self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
+
+    def test_normalized(self, mock_logger):
+        self.normalized = True
+        dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+            self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
+
+    def test_levenberg_marquardt(self, mock_logger):
+        self.opti_opts.search_method = 'levenberg_marquardt'
+        dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+            self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
+
+    def test_bad_method(self, mock_logger):
+        self.opti_opts.search_method = 'bad_method'
+        with self.assertRaises(ValueError):
+            dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+                self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
+
+    def test_minimums(self, mock_logger):
+        self.opti_opts.params[0].min_ = 10
+        dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+            self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
+
+    def test_huge_trust_radius(self, mock_logger):
+        # TODO: figure out how to get this to shrink a Newton step.
+        self.opti_opts.trust_radius = 1000000
+        dcs.bpe._dogleg_search(self.opti_opts, self.opti_opts.model_args, self.bpe_results, self.cur_results, \
+            self.delta_param, self.jacobian, self.gradient, self.hessian, normalized=self.normalized)
 
 #%% _analyze_results
-pass
+@patch('dstauffman.bpe.logger')
+class Test__analyze_results(unittest.TestCase):
+    r"""
+    Tests the _analyze_results function with the following cases:
+        Nominal
+        Normalized
+    """
+    def setUp(self):
+        self.opti_opts = dcs.OptiOpts()
+        self.opti_opts.params = [dcs.OptiParam('a'), dcs.OptiParam('b')]
+        self.bpe_results = dcs.BpeResults()
+        self.bpe_results.param_names = [x.encode('utf-8') for x in ['a', 'b']]
+        self.jacobian = np.array([[1, 2], [3, 4], [5, 6]])
+        self.normalized = False
+
+    def test_nominal(self, mock_logger):
+        dcs.bpe._analyze_results(self.opti_opts, self.bpe_results, self.jacobian, self.normalized)
+
+    def test_normalized(self, mock_logger):
+        self.normalized = True
+        dcs.bpe._analyze_results(self.opti_opts, self.bpe_results, self.jacobian, self.normalized)
+
+    def test_no_iters(self, mock_logger):
+        self.opti_opts.max_iters = 0
+        dcs.bpe._analyze_results(self.opti_opts, self.bpe_results, self.jacobian, self.normalized)
 
 #%% validate_opti_opts
 @patch('dstauffman.bpe.logger')
@@ -760,16 +867,35 @@ class Test_run_bpe(unittest.TestCase):
     def test_saving(self, mock_logger):
         mock_logger.setLevel(logging.CRITICAL)
         mock_logger.level = logging.CRITICAL
-        self.opti_opts.max_iters = 0
+        self.opti_opts.max_iters = 1
         self.opti_opts.output_folder = dcs.get_tests_dir()
         self.opti_opts.output_results = 'temp_results.hdf5'
         dcs.run_bpe(self.opti_opts)
         # TODO: test with more iterations and files?
 
+    def test_startup_finish_funcs(self, mock_logger):
+        self.opti_opts.start_func = lambda sim_params: dict()
+        self.opti_opts.final_func = lambda sim_params, settings: None
+        dcs.bpe.logger.setLevel(logging.CRITICAL)
+        mock_logger.level = logging.CRITICAL
+        dcs.run_bpe(self.opti_opts)
+
+    def test_failed(self, mock_logger):
+        # TODO: this case doesn't fail yet.  Make it do so.
+        self.opti_opts.max_iters = 1
+        self.opti_opts.tol_delta_step = 100
+        self.opti_opts.step_limit = 1
+        dcs.bpe.logger.setLevel(logging.CRITICAL)
+        mock_logger.level = logging.CRITICAL
+        dcs.run_bpe(self.opti_opts)
+
     def tearDown(self):
-        filename = os.path.join(self.opti_opts.output_folder, self.opti_opts.output_results)
-        if os.path.isfile(filename):
-            os.remove(filename)
+        if self.opti_opts.output_results:
+            files = [self.opti_opts.output_results, 'bpe_results_iter_1.hdf5', 'cur_results_iter_1.hdf5']
+            for this_file in files:
+                filename = os.path.join(self.opti_opts.output_folder, this_file)
+                if os.path.isfile(filename):
+                    os.remove(filename)
 
 #%% plot_bpe_results
 class Test_plot_bpe_results(unittest.TestCase):
